Move ase parallax sample to korge-sandbox
