+++ korge-sandbox/src/commonMain/resources/parallax_background.ase
