--- samples/parallax-scrolling-aseprite/build.gradle
@@ -1,4 +1,0 @@
-dependencies {
-	add("commonMainApi", project(":korge"))
-}
-
